deepen: absorb Summary Shell Metadata into its one caller
Replaces the shallow `getSummaryShellMetadata` module with a private
`fetchShellMetadata` helper inside the Server Component that renders
the summary-detail-page shell (`app/summary/[id]/page.tsx`). The
module had exactly one caller and its extracted interface was nearly
as complex as its implementation:

  export type SummaryShellMetadata = { ... };
  export async function getSummaryShellMetadata(id: string): Promise<SummaryShellMetadata> { ... }

Everything the module exported (the shape type + the async fetcher)
lived only to be re-imported by one page — a textbook extracted-for-
testability seam per LANGUAGE.md "one adapter means a hypothetical
seam". No tests referenced the module.

Dependency category: local-substitutable (per
process/Deepening/deepening.md §2). Prisma is a local-substitutable
adapter; the shell fetch is now a private helper inside the page
Server Component, tested at the page interface through the existing
`__tests__/app/summary/summary-page.test.tsx` harness (which mocks
`checkSummaryAccess` and never depended on the shell module).

Leverage: same as before — the page still flushes the shell before
Suspense'ing on `<SummaryBody>`, so the ~50ms warm shell query
survives unchanged. Callers gain nothing new here because there were
no other callers; maintainers gain locality — the auth-boundary docs
("NOT auth-checked here; middleware.ts:380 already gates
/summary/*") and the field selection now sit next to the JSX that
renders them, rather than three imports away.

Locality: a maintainer changing what the shell displays no longer
has to jump between `lib/db/get-summary-shell-metadata.ts` and
`app/summary/[id]/page.tsx` — the shape, the query, and the JSX
that consumes them live in one file.

Tests deleted: none (module had no tests). Tests added: none — the
existing skipped-plus-smoke suite in
`__tests__/app/summary/summary-page.test.tsx` continues to pass
unchanged (1 pass, 4 skip — the skipped tests reference
`checkSummaryAccess` at the SummaryBody boundary, not the shell).

Follows the same shape as ADR-0002 (Inline Analysis-Depth Scoring
into Onboarding Delivery): a pure-ish extracted-for-testability
function with one caller absorbed into that caller.
